fastapi-app/pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "logfire[fastapi,system-metrics]>=4.8.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.25",
]
